% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_app.R
\name{custom_app}
\alias{custom_app}
\title{custom app for generating subgraphs}
\usage{
custom_app(df_network = "random", directed = FALSE, coloring = TRUE)
}
\arguments{
\item{df_network}{dataframe with atleast two columns (from, to, weight)}

\item{directed}{Logical (default FALSE)}

\item{coloring}{node colouring based on edge-betweeness-algorithm}
}
\description{
custom app for generating subgraphs
}
